[feat/#267] 테스트 랜덤 로그인 버튼 삭제
