+++ src/hooks/auth/useSignIn.ts
@@ -1,7 +1,9 @@
 import authAxios from '@/lib/api/authAxios'
+import { TEST_ACOUNT } from '@/lib/data/constants'
 import { SIGN_IN_ERROR_MESSAGE } from '@/lib/data/messages'
 import { useUserStore } from '@/lib/stores/userStore'
 import { AuthResponse, SignInValues, User } from '@/lib/types/userTypes'
+import { getRandomInt } from '@/lib/utils/acountGenerator'
 import handleError from '@/lib/utils/errorHandler'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { AxiosError, AxiosResponse } from 'axios'
@@ -11,7 +13,17 @@ export default function useSignIn() {
   const queryClient = useQueryClient()
   const { setUser } = useUserStore()
 
-  return useMutation({
+  const randomSignIn = async (role: any) => {
+    const userList = TEST_ACOUNT[role]
+    const user = userList[getRandomInt(userList.length)]
+    console.log(`${role === 'APPLICANT' ? '지원자' : '사장님'} 테스트 계졍: `, {
+      email: user.email,
+      password: user.password,
+    })
+    await signIn.mutateAsync({ email: user.email, password: user.password })
+  }
+
+  const signIn = useMutation({
     mutationFn: async ({ email, password }: SignInValues): Promise<User> => {
       const response: AxiosResponse<AuthResponse> = await authAxios.post(
         '/auth/sign-in',
@@ -42,4 +54,6 @@ export default function useSignIn() {
       handleError(error, SIGN_IN_ERROR_MESSAGE)
     },
   })
+
+  return { signIn, randomSignIn }
 }
